packages/mtg-spellbook/pyproject.toml
[project]
name = "mtg-spellbook"
version = "0.1.0"
description = "MTG Spellbook - Interactive terminal UI for Magic: The Gathering card lookup and synergy exploration"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Vinnie Giarrusso", email = "vin.giarrusso@gmail.com" }]

dependencies = [
    "mtg-core",
    "textual>=0.80",
    "textual-image>=0.8",
    "pillow>=10.0,<13",
    "rich>=13.0,<15",
    "httpx>=0.27,<1",
    "ijson>=3.2,<4",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-textual-snapshot>=1.0.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
mtg-spellbook = "mtg_spellbook:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mtg_spellbook"]

[tool.uv.sources]
mtg-core = { workspace = true }
